adding Result<T> type to Activity

+++ application/Activities/Delete.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using application.Core;
 using domain;
 using MediatR;
 using persistence;
@@ -9,7 +10,7 @@ namespace application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest 
+        public class Command : IRequest<Result<Unit>>
         { 
             public Command(Guid id)
             {
@@ -19,7 +20,7 @@ namespace application.Activities
             public Guid Id { get; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext context;
 
@@ -28,15 +29,17 @@ namespace application.Activities
                 this.context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await context.Activities.FindAsync(new object[] { request.Id}, cancellationToken);
 
+                if (activity == null) return null;
+                
                 context.Activities.Remove(activity);
 
                 await context.SaveChangesAsync(cancellationToken);
 
-                return Unit.Value;
+                return Result.Success(Unit.Value);
             }
         }
     }
